Deleted minimizer plugin because it's already included in webpack4. Added global this value to this in library, so there is not necessary neither web nor node outputs.
